ODY-UI-01-002: real Unity test run + close out task, fix 3 bugs it found
The SQLite/Unity compile blocker (ODY-UI-01-002a, PR #68) is now merged.
Ran scripts/test-unity.ps1 for real on this branch and found three
independent, previously-undetectable issues -- the project had never
compiled/run in real Unity Editor before this point:

- BoardScreenPresenter.cs: TokenSizePixels (const double) assigned
  directly to StyleLength-typed style.width/.height -- CS0266, no
  implicit double->StyleLength conversion. Fixed with an explicit float
  cast (no behavioral change).
- BoardScreenPresenterTests.cs: missing `using Odyssey.Application.Commands;`
  for CommandId -- CS0246. Added the using directive.
- BoardScreenPresenterTests.cs: TemporaryDirectory.Dispose() intermittently
  threw IOException deleting campaign.db -- a Windows SQLite
  connection-pool file-locking race between the last SqliteConnection's
  Dispose() and directory cleanup (same class as ODY-UI-01-002a's own
  throwaway verification test hit). Fixed with a retry loop; no assertion
  or presenter-logic change.

Final scripts/test-unity.ps1 run: 0 compile errors, 40/40 EditMode
(including all 4 BoardScreenPresenterTests by name), 2/2 PlayMode.
dotnet build/test: 262/262, no regression. verify-format.ps1 and
check-repository-policy.ps1 both pass.

Updated ODY-UI-01-002's task contract (Status -> Done, AC-1-5/AC-8
Blocked -> Passed with real evidence) and SLICE-UI-01_IMPLEMENTATION_BACKLOG
row 1 (In Review -> Done). Historical blocker-discovery notes in the task
contract kept as honest history, not rewritten.

+++ Assets/Odyssey/Client/Runtime/BoardScreenPresenter.cs
@@ -142,8 +142,8 @@ namespace Odyssey.Unity.Client
                 VisualElement tokenElement = new VisualElement { name = "token-" + token.TokenId };
                 tokenElement.AddToClassList("board-token");
                 tokenElement.style.position = Position.Absolute;
-                tokenElement.style.width = TokenSizePixels;
-                tokenElement.style.height = TokenSizePixels;
+                tokenElement.style.width = (float)TokenSizePixels;
+                tokenElement.style.height = (float)TokenSizePixels;
                 tokenElement.style.left = ToPixels(token.Position.X);
                 tokenElement.style.top = ToPixels(token.Position.Y);
                 tokenElement.style.backgroundColor = new StyleColor(TokenColor(token, out _));
